Refactor

+++ database/migrations/2025_04_30_173428_create_telegraph_users_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('telegraph_users', function (Blueprint $table) {
                 $table->id();
-                $table->integer('user_id')->unique();
+                $table->bigInteger('user_id')->unique();
                 $table->string('token')->unique();
                 $table->timestamps();
         });
